Fixes trac:3557: Move openib init/finalize fix to v1.7
---svn-pre-commit-ignore-below---

r28317 [[BR]]
Be a little more conservative about initializing devices and modules
(i.e., ensure that more data items get zeroed out/set to NULL) so that
if something goes wrong during initialization, we don't try to clean
up something that isn't there (and segv).

The chance of this happening on the trunk is very low (and will also
be low once the verbs improvements are brought over to v1.7).  But it
can actually happen in the v1.6 branch (e.g., if no CPC is available,
we'll try to get the length of the endpoints list, but the endpoints
list is NULL).  

Hence, even though the real goal is to get this functionality over to
v1.6, I figured I'd commit to the trunk/CMR to v1.7 just to try to
keep commonality in the openib between all three where possible.

This commit was SVN r28478.

The following SVN revision numbers were found above:
  r28317 --> open-mpi/ompi@8405975bf62417e4aee83b1124ed339ed95c399d

The following Trac tickets were found above:
  Ticket 3557 --> https://svn.open-mpi.org/trac/ompi/ticket/3557