Fixes trac:3618: Move DDT fix to v1.6
---svn-pre-commit-ignore-below---

Custom patch from ticket based on:

r28553 [[BR]]
Patch based on Takahiro Kawashima fixing the issues with some
of the Fortran datatypes. This patch prevent the copy of the
datatype description from the OPAL to the OMPI layer in order
to decrease the memory requirements.

This commit was SVN r28585.

The following SVN revision numbers were found above:
  r28553 --> open-mpi/ompi@a9aae9c5383e225ff688c33266c2a51cb1bce548

The following Trac tickets were found above:
  Ticket 3618 --> https://svn.open-mpi.org/trac/ompi/ticket/3618