Fixes trac:3631: Move Intel Phi/openib fix to v1.7
---svn-pre-commit-ignore-below---

Custom patch on ticket based on:

r28589 [[BR]]
Refs trac:3626: that ticket specifically refers to the v1.6 branch; this
commit is the trunk version of what is needed for #3626.

Add the "ignore_device" field to the INI file.  This allows us to
specifically list devices that should be ignored by the openib BTL
(such as the Intel Phi, at least as of May 2013 -- see #3626).  

Also add the Intel Phi to the ini file, and set its ignore_device=1.

Finally, add the concept of counting intentionally ignored verbs
devices.  Devices are ignored for one of two reasons:

 * If the number of allowed ports on that device is 0 (i.e., if
   if_include/if_exclude was set such that we're intentionally
   ignoring this device).
 * If the INI ignore_device field for this device is set to 1.

Once we have the count of devices that were intentionally ignored,
only show the "Hey, there's verbs devices that you're not using!"
show_help message if there are devices that were ''unintentionally''
ignored.

This commit was SVN r28625.

The following SVN revision numbers were found above:
  r28589 --> open-mpi/ompi@713e3aa3dbe3e2ddd351da0bc86776c56cc83f4d

The following Trac tickets were found above:
  Ticket 3626 --> https://svn.open-mpi.org/trac/ompi/ticket/3626
  Ticket 3631 --> https://svn.open-mpi.org/trac/ompi/ticket/3631