Fixes trac:3642: Move r28658 to v1.7 branch
---svn-pre-commit-ignore-below---

r28658 [[BR]]
Per an off-list discussion, it appears possible for a system to report failure when executing getpwuid. There are several reasons for this error to occur, most notably if the system uses a network-based authentication protocol (e.g., NIS) and that sytem gets overwhelmed when we launch on a lot of nodes.

There is no good way to recover from this scenario, and from past experience, using the user's name in the session directory (as opposed to the uid) is very helpful when things go wrong. So print a help message when this happens (it is extremely rare, but has happened at least once now) and return an error.

cmr:v1.7.3,reviewer=jsquyres
cmr:v1.6.5,reviewer=jsquyres


r28662 [[BR]]
Tweaked the help message a bit (this is the end result of iterating on
the message in email between Mike, Ralph, Jeff).

Add this to CMR #3642 and #3643.

This commit was SVN r28664.

The following SVN revision numbers were found above:
  r28658 --> open-mpi/ompi@13665bffe85bd35a17b14072cd46a127d6cb6a74
  r28662 --> open-mpi/ompi@b9ca8e3cd1f9aabf8eb5260e6b6a59baac700981

The following Trac tickets were found above:
  Ticket 3642 --> https://svn.open-mpi.org/trac/ompi/ticket/3642