Fixes trac:3673: Move r28790 to v1.7 branch (datatype packing/unpacking fix)
---svn-pre-commit-ignore-below---

r28790 [[BR]]
Fix an issue with the packing/unpacking of datatype representations
identified by Takahiro Kawashima. The packed length was reported as a
max bound and not provided on the unpacking side, so the unpacking
buffer could become out of sync with the content stored after the
packed representation.

The fix force the packing operation itself before reporting the length,
so we always report now the real number of bytes in the packed
representation.

cmr:v1.7.3:reviewer=jsquyres

This commit was SVN r28807.

The following SVN revision numbers were found above:
  r28790 --> open-mpi/ompi@f7ac610bee89595d8692fbede7276f2ef603e59f

The following Trac tickets were found above:
  Ticket 3673 --> https://svn.open-mpi.org/trac/ompi/ticket/3673