Fixes trac:3671: Move r28789 to v1.7 (remove OPAL_TRACE)
---svn-pre-commit-ignore-below---

r28789 [[BR]]
Turns out that there was exactly ONE place in all of the OMPI code base that still referred to OPAL_TRACE, though a few places retained the include file for no reason. So no point in letting this sit as it is clearly an unused "feature".

This commit was SVN r28808.

The following SVN revision numbers were found above:
  r28789 --> open-mpi/ompi@10ca1c1b048974a08ef36951790f2e2196f180f1

The following Trac tickets were found above:
  Ticket 3671 --> https://svn.open-mpi.org/trac/ompi/ticket/3671