Fixes trac:3676: Move opal/if and hotel improvements to v1.7
---svn-pre-commit-ignore-below---

r28343 [[BR]]
Defensive programming: add a constructor for opal_if_t that zeros
everything out before using it.  

This is not in response to any known bug, but rather just a
pre-emptive, defensive move to help prevent bugs in code that forgets
to initialize a field.

r28344 [[BR]]
Also support getting the MAC and MTU.

r28345 [[BR]]
Add public functions for retrieving the MAC and MTU (paired with
r28344).

r28395 [[BR]]
Ahem....cleanup a copy/paste error in naming of these functions

r28757 [[BR]]
Add an oh-so-slightly faster variant of the hotel "checkin" action
(since this is used in the fast path) for when you ''know'' that there
will be a room available:

 * Don't do the last_unoccupied_room check
 * Return void

This commit was SVN r28827.

The following SVN revision numbers were found above:
  r28343 --> open-mpi/ompi@eb012c2aad707bfe1bc872da0cf1cc632b4d2039
  r28344 --> open-mpi/ompi@e88881c25f1d0b008e936c6bfbaa24b96b11afd5
  r28345 --> open-mpi/ompi@c722440411f078bac5606a79a91d1c78f194e1b8
  r28395 --> open-mpi/ompi@cef639f5785324e3410085998b744ac108e44e39
  r28757 --> open-mpi/ompi@bdb45a2e4f406cc419ab4c7e3e032dc15898c18a

The following Trac tickets were found above:
  Ticket 3676 --> https://svn.open-mpi.org/trac/ompi/ticket/3676