Fixes trac:3747: Large memory consuption for XRC mode in OpenIB BTL
We investigated memory consumption by XRC mode in Open IB BTL and found it is too large. It happens because the bug in the function mca_btl_openib_size_queues() where the length for send and recv CQ is calcualted as for Peer-to-Peer endpoints. Attached file openib.c.diff is a patch for given problem. And we suggest the patch btl_openib_endoint.c.diff that reduces memory consuption for XRC QP in case of great PPN. 

Provided by aryzhikh as patch.

This commit was SVN r29172.

The following Trac tickets were found above:
  Ticket 3747 --> https://svn.open-mpi.org/trac/ompi/ticket/3747