Fixes trac:3775: Move r29201 to v1.7 branch (Silence a warning about an)
---svn-pre-commit-ignore-below---

r29201 [[BR]]
Silence a warning about an address that can never be NULL - libevent needs to deal with the situation where the user may have compiled the code on a system where this function is present, but executes it on one where it isn't. Thus, a compile-time test isn't adequate.

Pushed upstream.

cmr:v1.7.3:reviewer=jsquyres


r29210 [[BR]]
C standard stipulates that we have to cast the function to another of the same type to avoid unexpected behavior. We aren't using the function in this case, but Nick correctly points out that we should follow the standard regardless.

Refs trac:3755

This commit was SVN r29220.

The following SVN revision numbers were found above:
  r29201 --> open-mpi/ompi@7de493fc023047faf084a3c4ff4303dfcd61407a
  r29210 --> open-mpi/ompi@7bc20866fd184bb46908b55138e5582b9c8fe832

The following Trac tickets were found above:
  Ticket 3755 --> https://svn.open-mpi.org/trac/ompi/ticket/3755
  Ticket 3775 --> https://svn.open-mpi.org/trac/ompi/ticket/3775