Allow the trunk to cimpile after r29265. The addition of the neighborhood collective to the mca_coll_base_comm_coll_t structure increased the size of the ompi_communicator_t over the limit of the predefined padding (PREDEFINED_COMMUNICATOR_PAD).
This fix is a temporary fix to allow the trunk to compile. Unfortuantely
it breaks the compatibility with all other versions of Open MPI. Please
read the comment in this header file for a more complete explanation.

This commit was SVN r29277.

The following SVN revision numbers were found above:
  r29265 --> open-mpi/ompi@c5596548b2eb8fba40bdff48b4f002a9619ebbe9