Fixes trac:3814: Move r29298 to v1.7 branch (fix [i]alltoallw mpi_f08 wrapper functions)
---svn-pre-commit-ignore-below---

r29298 [[BR]]
Make the Alltoall/Ialltoallw F08 interfaces a little less sucky.

Use proper array types for the sendtypes and recevtypes params in the
definition of these subroutines (since they're prototyped to be array
parameters in the module).  See the comment in the code for a more
complete explanation.

Note that after talking to Craig Rasumussen, he says that this is
fairly common practice in the Fortran community (i.e., pass a scalar
that is part of an array to effectively pass an offset into that
array, since Fortran passes by reference).  So this might be a hack,
but it's a commonly-accepted hack.

Reviewed by Craig Rasmussen.

cmr=v1.7.3:subject=Fix [I]Alltoallw mpi_f08 wrapper functions

r29301 [[BR]]
Missed updating the profiling versions of [I]Alltoallw in r29298.

Refs trac:3814.

This commit was SVN r29309.

The following SVN revision numbers were found above:
  r29298 --> open-mpi/ompi@0acebc0a1f9145bcb182e385bf9a321a2665ae41
  r29301 --> open-mpi/ompi@4af194ca32ae51b4cb154d35ef47f6db89ecfd04

The following Trac tickets were found above:
  Ticket 3814 --> https://svn.open-mpi.org/trac/ompi/ticket/3814