Fixes trac:3811: Move r29285 to v1.7 branch (fix convertor set_position)
---svn-pre-commit-ignore-below---

r29285 [[BR]]
Fix a corner case for a non-contiguous send convertor where the
convertor accepted to be set to a position in the middle of a
predefined datatype. Once set there is was unable to provide the
second part of the datatype. This fix force the convertor to be
aligned on predefined datatypes boundaries for any non-contiguous
send convertor.

r29296 [[BR]]
regression test for r29285 (convertor_set_position)

This commit was SVN r29310.

The following SVN revision numbers were found above:
  r29285 --> open-mpi/ompi@43b4d76913ec3478775341cc17a2cbec2d78f78a
  r29296 --> open-mpi/ompi@002ba95deb8f4149e97524d1f434564f3fc21dec

The following Trac tickets were found above:
  Ticket 3811 --> https://svn.open-mpi.org/trac/ompi/ticket/3811