add a .mailmap file for the github mirror
This file exists to help map usernames to proper names and email
addresses in the Open MPI github mirror of the canonical SVN repository.
The github mirror can be found here:

  https://github.com/open-mpi/ompi-svn-mirror

I've seeded the file with the names of Cisco contributors.  In order to
avoid exposing anyone's email address without their permission, we are
using an opt-in model for adding real email addresses.