Fixes trac:3859 Rollup remaining ORTE changes to sync v1.7 to trunk
Contains a wide range of CMRs (logged into the svn:mergeinfo property) that include:
  * a rewrite of ORTE's oob/rml subsystems to support async messaging
  * add dynamic allocation support
  * modify singleton support so it no longer immediately spawns an HNP
  * support for MapReduce operations

and a variety of other features that I frankly have forgotten.

This is not undergoing a formal review as it is too large and widespread for that purpose. So we'll let this soak for some time, fixing errors that appear via MTT on various systems.

This commit was SVN r29587.

The following Trac tickets were found above:
  Ticket 3859 --> https://svn.open-mpi.org/trac/ompi/ticket/3859