Fixes trac:3847: Move r29435, 29438, 29439, 29475, r29504 to v1.7 branch
---svn-pre-commit-ignore-below---

r29435 [[BR]]
****************************************************************
       This change contains a non-mandatory modification
       of the MPI-RTE interface. Anyone wishing to support
       coprocessors such as the Xeon Phi may wish to add
       the required definition and underlying support
****************************************************************

Add locality support for coprocessors such as the Intel Xeon Phi.

Detecting that we are on a coprocessor inside of a host node isn't straightforward. There are no good "hooks" provided for programmatically detecting that "we are on a coprocessor running its own OS", and the ORTE daemon just thinks it is on another node. However, in order to properly use the Phi's public interface for MPI transport, it is necessary that the daemon detect that it is colocated with procs on the host.

So we have to split the locality to separately record "on the same host" vs "on the same board". We already have the board-level locality flag, but not quite enough flexibility to handle this use-case. Thus, do the following:

1. add OPAL_PROC_ON_HOST flag to indicate we share a host, but not necessarily the same board

2. modify OPAL_PROC_ON_NODE to indicate we share both a host AND the same board. Note that we have to modify the OPAL_PROC_ON_LOCAL_NODE macro to explicitly check both conditions

3. add support in opal/mca/hwloc/base/hwloc_base_util.c for the host to check for coprocessors, and for daemons to check to see if they are on a coprocessor. The former is done via hwloc, but support for the latter is not yet provided by hwloc. So the code for detecting we are on a coprocessor currently is Xeon Phi specific - hopefully, we will find more generic methods in the future.

4. modify the orted and the hnp startup so they check for coprocessors and to see if they are on a coprocessor, and have the orteds pass that info back in their callback message. Automatically detect that coprocessors have been found and identify which coprocessors are on which hosts. Note that this algo isn't scalable at the moment - this will hopefully be improved over time.

5. modify the ompi proc locality detection function to look for coprocessor host info IF the OMPI_RTE_HOST_ID database key has been defined. RTE's that choose not to provide this support do not have to do anything - the associated code will simply be ignored.

6. include some cleanup of the hwloc open/close code so it conforms to how we did things in other frameworks (e.g., having a single "frame" file instead of open/close). Also, fix the locality flags - e.g., being on the same node means you must also be on the same cluster/cu, so ensure those flags are also set.

cmr:v1.7.4:reviewer=hjelmn


r29438 [[BR]]
Correct location of elog file

Refs trac:3847


r29439 [[BR]]
Per a good suggestion from Jeff, make the coprocessor mapping more scalable by using a hash table to cache the coprocessor list, and then do a single pass thru the nodes at the end to assign hostid's.

Refs trac:3847


r29475 [[BR]]
Fix locality when in the case where the OMPI_RTE_HOST_ID is not found.

cmr=v1.7.4:ticket=3847

r29489 [[BR]]
Fix coprocessor detection by always adding the local daemon's co-processors
to the hash table.

Tested and working on a system with 2 Xeon Phi co-processors.

cmr=v1.7.4:ticket=3847:reviewer=ompi-rm1.7

r29504 [[BR]]
Initialize orte_coprocessors hash table to NULL. Delay coprocessor detection on HNP until after node topology final definition in case rmaps changes it. Minor spacing change.

Refs trac:3847

This commit was SVN r29603.

The following SVN revision numbers were found above:
  r29435 --> open-mpi/ompi@24c811805f14a27e1f936dffe49f490708d6b5a6
  r29438 --> open-mpi/ompi@772a376d73f5913c0173db8322fcbcf1849d361e
  r29439 --> open-mpi/ompi@b12167abef09ea8938ddea8b97bd97c7b793a258
  r29475 --> open-mpi/ompi@5bf65556046f72c80da1b6441cc19ded0f2bfc8a
  r29489 --> open-mpi/ompi@2e2794fa15eb2a21bb7b98b03fbb0c02a6b4ae59
  r29504 --> open-mpi/ompi@604970a1a252401a88a1eb3f02b17f4aca344041

The following Trac tickets were found above:
  Ticket 3847 --> https://svn.open-mpi.org/trac/ompi/ticket/3847