Fixes trac:3932: Prevent extent retrieval on the send datatype on non-root processes for MPI_Iscatter
---svn-pre-commit-ignore-below---

r29736 [[BR]]
Fix an error message for the igather and the usage of the extent on
non non-root processes for the iscatter. Thanks to Pierre
Jolivet for the bug report and the patch.

r29761 [[BR]]
Restrict the usage of MPI_Type_extent only to receiving processes
(aka the root). This commit is based on a patch provided by Pierre 
Jolivet.
Fix all the output to match the failing MPI call.

This commit was SVN r29768.

The following SVN revision numbers were found above:
  r29736 --> open-mpi/ompi@68268377afc6220618227b81bfc1273f125f19ee
  r29761 --> open-mpi/ompi@cb2427773783db6e8e79f8332ccb3243d1ecab03

The following Trac tickets were found above:
  Ticket 3932 --> https://svn.open-mpi.org/trac/ompi/ticket/3932