Fixes trac:3966: Move PORT_ACTIVE usnic fix to v1.7
---svn-pre-commit-ignore-below---

r29852 [[BR]]
Technically, the PORT_ACTIVE is not a "bad" event.

Note that this event should never happen within a single OMPI job,
because OMPI will ignore usnic ports that are down.  The PORT_ACTIVE
event should only occur if a port ''was'' down and is now ''up''.  But
what the heck -- if we ever do get this event, it is harmless -- just
ignore it.

This commit was SVN r29863.

The following SVN revision numbers were found above:
  r29852 --> open-mpi/ompi@0f61bb651e40b3f744dbe314e4281f8d194a279a

The following Trac tickets were found above:
  Ticket 3966 --> https://svn.open-mpi.org/trac/ompi/ticket/3966