Fixes trac:3940: Update udcm support in v1.7 branch
---svn-pre-commit-ignore-below---

r29767 [[BR]]
openib/connect: re-enable xrc support in the openib btl

This commit updates the udcm cpc to support xrc. The steps followed by udcm
mimic those in the removed xoob cpc. This update has been tested with both XRC
and RC.

Mellanox, this is intended to go into 1.7.4. Please review carefully and let
me know if there are any issues.

cmr=v1.7.4:reviewer=miked

r29775 [[BR]]
udcm: cleanup code and improve the ack handling

Originally udcm acks used the immediate data to indicate which message was
being acknowleged. This data was (mysteriously) junk when using QLogic HCAs so I
updated udcm to use the source info (slid, qp, etc) to determine which message was being
acked. This works as long as we don't have two messages simultaneously in flight
to a particular peer and then loose the first of the two messages. The chances of this
happening are tiny. To fix this case I updated the udcm message header to include
a pointer to the in flight message. This pointer is then sent back to the sending
process to ack receipt.

cmr=v1.7.4:ticket=trac:3940

r29859 [[BR]]
Fix warnings in udcm.

cmr=v1.7.4:reviewer=rhc:ticket=3940


r29861 [[BR]]
Cleanup udcm structures to avoid issues with nesting structures with
flexible members.

UDCM is ready to go for 1.7.4 with this patch.

cmr=v1.7.4:ticket=3940

This commit was SVN r29870.

The following SVN revision numbers were found above:
  r29767 --> open-mpi/ompi@fb0b0442c40a4080b22d189c967837357e3c8eb7
  r29775 --> open-mpi/ompi@fe327d985931eec086fe7fe29384019d66979767
  r29859 --> open-mpi/ompi@6ab69c758b74a08544758664b2381b71f91bf869
  r29861 --> open-mpi/ompi@3262080391a2bb70ce3f99947fe20d0b9923a2cf

The following Trac tickets were found above:
  Ticket 3940 --> https://svn.open-mpi.org/trac/ompi/ticket/3940