Fixes trac:3952: Add minor configure fix to v1.7
---svn-pre-commit-ignore-below---

r29801 [[BR]]
Ensure that ompi_fortran_happy is set to 0 when we're not building the
OMPI project.  This is necessary because some of the threading tests
(that are even used in ORTE) use this result.

This commit was SVN r29888.

The following SVN revision numbers were found above:
  r29801 --> open-mpi/ompi@a2aee3ecf107dd9e9a691c9dfe192a3399a74b15

The following Trac tickets were found above:
  Ticket 3952 --> https://svn.open-mpi.org/trac/ompi/ticket/3952