Fixes #3989: opal_pack_debug was instantiated as a bool in one file and extern'ed s an int in another.  This caused a SIBGUS on Solaris/SPARC.
This commit properly moves the extern to a .h file so that it's the
same in all files.  It also moves the DO_DEBUG to the header file,
because it was defined to the same thing in multiple .c files.

cmr=v1.7.4:reviewer=bosilca:subject=fix SPARC SIGBUS in opal convertor code