Fixes trac:3999: Move r29978 to v1.7 branch (eliminate excessive binding/memory warnings)
---svn-pre-commit-ignore-below---

r29978 [[BR]]
Don't warn about binding (process and/or memory) if the node cannot do it or if we would overload, but it wasn't specifically requested by the user (i.e., it is the result of the default policy). Instead, just don't bind and quietly move along.

Reset topology usage for each node as we bind as multiple nodes may be linked to the same topology object. This will need to be revisited for scale as it does take some non-zero time to reset the usage each iteration. However, storing individual topology objects for every node consumes memory, so it's a tradeoff.

cmr=v1.7.4:reviewer=jsquyres:subject=Eliminate excessive binding/memory warnings

This commit was SVN r30017.

The following SVN revision numbers were found above:
  r29978 --> open-mpi/ompi@55cd65b1491c34080391f699d8e0e5812b2dbe7d

The following Trac tickets were found above:
  Ticket 3999 --> https://svn.open-mpi.org/trac/ompi/ticket/3999