Fixes trac:4004: Move r29991 to v1.7 branch (fix sparc sigbus in opal convertor code)
---svn-pre-commit-ignore-below---

r29991 [[BR]]
Fixes trac:3989: opal_pack_debug was instantiated as a bool in one file
and extern'ed s an int in another.  This caused a SIBGUS on
Solaris/SPARC.

This commit properly moves the extern to a .h file so that it's the
same in all files.  It also moves the DO_DEBUG to the header file,
because it was defined to the same thing in multiple .c files.

cmr=v1.7.4:reviewer=bosilca:subject=fix SPARC SIGBUS in opal convertor code

r30018 [[BR]]
Cleanup all the datatype test to avoid any memory leaks or RUI from valgrind.

r30019 [[BR]]
Correctly initialize and finalize all the datatype classes. No memory leaks on the
datatype engine subsists.

r30020 [[BR]]
rearrange the fields to remove a gap in the datatype.

r30021 [[BR]]
Remove unused variable

Refs trac:4004

This commit was SVN r30022.

The following SVN revision numbers were found above:
  r29991 --> open-mpi/ompi@a88e143127fc40bbcc65b640b76b292651287b23
  r30018 --> open-mpi/ompi@a85194ae96ec7dc185432a91ab8436dadda7d2d5
  r30019 --> open-mpi/ompi@7178492dd5cd3737423db37a086fe54aa0477fc6
  r30020 --> open-mpi/ompi@178c34099252d25ec39dc19893edeb2651c31871
  r30021 --> open-mpi/ompi@f026bdb68b8edc36392219cdabcf668b4870e487

The following Trac tickets were found above:
  Ticket 3989 --> https://svn.open-mpi.org/trac/ompi/ticket/3989
  Ticket 4004 --> https://svn.open-mpi.org/trac/ompi/ticket/4004