Fixes trac:4019: Move r30042 to v1.7 branch (remove an annoying warning)
---svn-pre-commit-ignore-below---

r30042 [[BR]]
Remove an annoying warning. If the user excludes a non-existent interface, there is no reason to warn - the interface may simply not exist on that node.

cmr=v1.7.4:reviewer=jsquyres:subject=Remove an annoying warning


r30091 [[BR]]
As Jeff pointed out, the reqd flag should only turn off the show_help - still enter the rest of the code block

Refs trac:4019


r30146 [[BR]]
Per 1/7/2014 telecon: Add an MCA param to turn on all warnings for missing excluded interfaces.

Refs trac:4019


r30161 [[BR]]
Change the MCA level to 2 (from 5) on the rationale that it may be
needed for correctness.  The if_include/if_exclude are level 1, and
the TCP port range params are level 2; this parameter seems to be on
par with the TCP port range params.

Refs trac:4019

This commit was SVN r30165.

The following SVN revision numbers were found above:
  r30042 --> open-mpi/ompi@042ed95e4ea3fe350d0029ff4b01d36b297189d6
  r30091 --> open-mpi/ompi@62378a64c840aeae05d192459a2378497664ed96
  r30146 --> open-mpi/ompi@e2ca265f400933ab689427a5eeb028fcaf04abd8
  r30161 --> open-mpi/ompi@9d41632eba16f6a5423cf6f52bda424e39cdef28

The following Trac tickets were found above:
  Ticket 4019 --> https://svn.open-mpi.org/trac/ompi/ticket/4019