Addendum to #4079: Move r30198 to v1.7 branch (make statfs/statvfs more robust)
Problem was detected on OS X Mavericks after the original CMR was
completed, which led to r30206 on the trunk which fixes the problem.
RM gave me verbal approval to simply include this fix in the amended
CMR.

Refs trac:4079

---svn-pre-commit-ignore-below---

r30206 [[BR]]
Fix typo in macro name (#$%@#$% defined-or-not macros!!)

Refs trac:4079

This commit was SVN r30207.

The following SVN revision numbers were found above:
  r30198 --> open-mpi/ompi@c67c8e81878160dce22bd1b4bc644e635f1aaefd
  r30206 --> open-mpi/ompi@023c50e864e695d325f0ed07a83ea655842dc248

The following Trac tickets were found above:
  Ticket 4079 --> https://svn.open-mpi.org/trac/ompi/ticket/4079