Make the non-Linux platforms support an interactive opal_path_nfs() test.
On Linux, if this test is run with no command line params, it will run
"mount" and analyze the output (same as it always has).

On all platforms, if you provide one or more command line options,
each command line option is given to opal_path_nfs() and the result is
sent to stdout.

This commit was SVN r30208.